+++ smart-contracts/contracts/basic-vault/src/multitest/vault.rs
@@ -1,5 +1,6 @@
 use crate::multitest::common::*;
 use crate::multitest::suite::*;
+use quasar_types::error::Error::ItemIsEmpty;
 
 #[test]
 fn try_bond() {
@@ -16,8 +17,8 @@ fn try_bond() {
     // this error happens because our ibc channel is not open yet
     assert_eq!(
         err,
-        VaultContractError::Std(cosmwasm_std::StdError::GenericErr {
-            msg: "alloc::string::String not found".to_string()
+        VaultContractError::QuasarError(ItemIsEmpty {
+            item: "Item icq_channel is empty".to_string()
         })
     );
 }