+++ smart-contracts/packages/quasar-types/CHANGELOG.md
@@ -1,8 +1,19 @@
 # CHANGELOG
 
 ## Unreleased
+### Dependencies
+### API breaking
+### State breaking
+### Improvements
+- Removed quasar-traits folder as it was not used
+### Features
+- Created ItemShouldLoad trait and implemented it for Item to show namespace on error
+- Created MapShouldLoad trait and implemented it for Map to show namespace & key on error
+- Added tests for ITEM.should_load() & MAP.should_load()
+### Bugfixes
 
 ## V0.1.0
+
 <!-- ### Initial version
 ### Dependencies
 ### API breaking